[move] enum파일 domain/vo로 이동
