+++ company/src/main/java/com/hubEleven/company/presentation/controller/CompanyController.java
@@ -10,8 +10,8 @@ import com.hubEleven.company.application.command.SearchCompanyCommand;
 import com.hubEleven.company.application.command.UpdateCompanyCommand;
 import com.hubEleven.company.application.dto.response.CompanyResult;
 import com.hubEleven.company.application.service.CompanyAppService;
-import com.hubEleven.company.domain.model.CompanyStatus;
-import com.hubEleven.company.domain.model.CompanyType;
+import com.hubEleven.company.domain.vo.CompanyStatus;
+import com.hubEleven.company.domain.vo.CompanyType;
 import com.hubEleven.company.presentation.dto.request.CompanyRequests;
 import com.hubEleven.company.presentation.dto.response.CompanyResponse;
 import io.swagger.v3.oas.annotations.Operation;
